refactor: Participants 를 생성해 dealer 와 players 를 한번에 관리
