+++ src/main/java/blackjack/view/ResultView.java
@@ -1,40 +1,45 @@
 package blackjack.view;
 
-import blackjack.domain.BlackjackDto;
+import blackjack.domain.ParticipantsDto;
 import blackjack.domain.Participant;
+import blackjack.domain.Participants;
 
 public class ResultView {
 
     private static final String RESULT_DELIMITER = " - 결과: ";
 
-    public void showStartingStatus(BlackjackDto blackjackDto) {
-        System.out.println(blackjackDto.getDealerOpenCard());
-        for (Participant player : blackjackDto.getPlayers()) {
-            showEachPlayerCurrentStatus(blackjackDto, player);
+    public void showStartingStatus(ParticipantsDto participantsDto) {
+        System.out.println(participantsDto.getDealerOpenCard());
+        for (Participant player : participantsDto.getPlayers()) {
+            showEachPlayerCurrentStatus(participantsDto, player);
         }
     }
 
-    public void showCurrentStatus(BlackjackDto blackjackDto) {
-        Participant dealer = blackjackDto.getDealer();
-        showEachPlayerCurrentStatus(blackjackDto, dealer);
-        for (Participant player : blackjackDto.getPlayers()) {
-            showEachPlayerCurrentStatus(blackjackDto, player);
+    public void showCurrentStatus(ParticipantsDto participantsDto) {
+        Participant dealer = participantsDto.getDealer();
+        showEachPlayerCurrentStatus(participantsDto, dealer);
+        for (Participant player : participantsDto.getPlayers()) {
+            showEachPlayerCurrentStatus(participantsDto, player);
         }
     }
 
-    public void showEachPlayerCurrentStatus(BlackjackDto blackjackDto, Participant participant) {
-        System.out.println(getEachPlayerStatus(blackjackDto, participant));
+    public void showEachPlayerCurrentStatus(ParticipantsDto participantsDto, Participant participant) {
+        System.out.println(getEachPlayerStatus(participantsDto, participant));
     }
 
-    private String getEachPlayerStatus(BlackjackDto blackjackDto, Participant participant) {
-        return blackjackDto.getPlayerCardStatus(participant);
+    private String getEachPlayerStatus(ParticipantsDto participantsDto, Participant participant) {
+        return participantsDto.getPlayerCardStatus(participant);
     }
 
-    public void showFinishedStatus(BlackjackDto blackjackDto) {
-        Participant dealer = blackjackDto.getDealer();
-        System.out.println(getEachPlayerStatus(blackjackDto, dealer) + RESULT_DELIMITER + dealer.getScore());
-        for (Participant player : blackjackDto.getPlayers()) {
-            System.out.println(getEachPlayerStatus(blackjackDto, player) + RESULT_DELIMITER + player.getScore());
+    public void showFinishedStatus(ParticipantsDto participantsDto) {
+        Participant dealer = participantsDto.getDealer();
+        System.out.println(getEachPlayerStatus(participantsDto, dealer) + RESULT_DELIMITER + dealer.getScore());
+        for (Participant player : participantsDto.getPlayers()) {
+            System.out.println(getEachPlayerStatus(participantsDto, player) + RESULT_DELIMITER + player.getScore());
         }
     }
+
+	public void showResult(Participants participants) {
+
+	}
 }
