+++ src/test/java/blackjack/CardDeckTest.java
@@ -12,8 +12,7 @@ public class CardDeckTest {
 	@Test
 	@DisplayName("pickCard 에서 카드를 잘 뽑는지")
 	void PickCard_Return_Only_One() {
-		CardDeck cardDeck = new CardDeck();
-		assertThat(cardDeck.pickCard()).isInstanceOf(Card.class);
+		assertThat(CardDeck.pickCard()).isInstanceOf(Card.class);
 	}
 
 }
